Only delete root node to delete nested tree
Acccording to the [solr
docs](https://solr.apache.org/guide/solr/latest/indexing-guide/indexing-nested-documents.html#maintaining-integrity-with-updates-deletes-and-shard-splits)
we only need to delete the root node of the finding aid, by id, and all child documents will also be deleted.

This is much faster and probably ultimately better for internal
stability of our index.